server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
yfinance==0.2.18
pandas==2.0.3
numpy==1.25.2
requests==2.31.0
python-socketio==5.8.0
eventlet==0.33.3
PyJWT==2.8.0
pymongo==4.5.0
gunicorn==21.2.0